requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.4.1
torchvision==0.19.1+cu118
xformers==0.0.28
imageio==2.33.0
Pillow==10.0.1
imageio-ffmpeg==0.5.0
omegaconf==2.1.1
einops>= 0.4.1 
transformers==4.30.2
torchmetrics>= 0.11.0
datasets==2.14.5
rich>=13.3.2
wandb==0.18.7
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
ipdb
tqdm
easydict==1.13
lpips==0.1.4
jaxtyping==0.3.1
scikit-image==0.23.2 
numpy==1.26.4